Fix DISPLAY_EFFE, but also remove it from the render options

+++ src/graphics.c
@@ -2416,7 +2416,7 @@ void render_parts(pixel *vid)
 				{
 					gradv = 0.4f;
 					if (!(parts[i].life<5))
-						q = (int)sqrt(parts[i].life);
+						q = (int)sqrtf(parts[i].life);
 					else
 						q = parts[i].life;
 					colr = colg = colb = (int)(sin(gradv*q) * 100 + 128);
